requirements.txt
python-socketio==5.1.0
websocket-client==0.58.0